requirements.txt
openrgb-python>=0.2.0
Pillow>=9.0.0
sounddevice>=0.4.0
numpy>=1.21.0 